luga_Product/luga_Backend_master/requirements.txt
fastapi
pydantic
uvicorn
motor
pymongo
python-dotenv
httpx
requests
python-multipart
PyPDF2
python-docx
cloudinary
openai